+++ api/src/certification/results/domain/usecases/index.js
@@ -5,9 +5,11 @@ import { fileURLToPath } from 'node:url';
 import * as certificationResultRepository from '../../../../../lib/infrastructure/repositories/certification-result-repository.js';
 import { injectDependencies } from '../../../../shared/infrastructure/utils/dependency-injection.js';
 import { importNamedExportsFromDirectory } from '../../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
+import * as sessionEnrolmentRepository from '../../../enrolment/infrastructure/repositories/session-repository.js';
 import * as certificationCourseRepository from '../../../shared/infrastructure/repositories/certification-course-repository.js';
 import * as certificationReportRepository from '../../../shared/infrastructure/repositories/certification-report-repository.js';
 import * as certificateRepository from '../../infrastructure/repositories/certificate-repository.js';
+import * as cleaCertifiedCandidateRepository from '../../infrastructure/repositories/clea-certified-candidate-repository.js';
 import * as scoCertificationCandidateRepository from '../../infrastructure/repositories/sco-certification-candidate-repository.js';
 
 /**
@@ -18,13 +20,18 @@ import * as scoCertificationCandidateRepository from '../../infrastructure/repos
  * @typedef {certificationCourseRepository} CertificationCourseRepository
  * @typedef {certificateRepository} CertificateRepository
  * @typedef {certificationReportRepository} CertificationReportRepository
+ * @typedef {cleaCertifiedCandidateRepository} CleaCertifiedCandidateRepository
+ * @typedef {sessionEnrolmentRepository} SessionEnrolmentRepository
  **/
+
 const dependencies = {
   certificationCourseRepository,
   certificationResultRepository,
   scoCertificationCandidateRepository,
   certificateRepository,
   certificationReportRepository,
+  cleaCertifiedCandidateRepository,
+  sessionEnrolmentRepository,
 };
 
 const path = dirname(fileURLToPath(import.meta.url));
